Better project organization, abstracted out a significant amount of code
